fix: disable side-effects for no_predicates functions (#6027)
# Description

## Problem\*

Brillig output is nullified if the call is with a predicate, however,
functions with no_predicates attribute expect valid values from their
brillig calls because when they constrain the values , it will not have
a predicate.

Resolves e-2-e test failing on PR
https://github.com/AztecProtocol/aztec-packages/pull/4571

## Summary\*
Disable side-effects when calling a function with no_predicates



## Additional Context
The dedicated inlining pass for no_predicates functions can be removed
after this fix.


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.