feat: add `Expr::as_constructor` (#5980)
# Description

## Problem

Part of #5668

## Summary

Also add doc comments for `expr.nr`.

## Additional Context

I wasn't sure how to represent a `Path`... so I did it with a `Quoted`
value. I first thought about using a `Variable`, but then that wouldn't
be parsed right when parsing a constructor, so it had to be some kind of
Path.

## Documentation

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>