fix: Parse a statement as an expression (#6040)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/6036

## Summary\*

Grego was having an issue from some macros where some code was taken
from a block and later put in the rhs of a `let` statement: `let _ =
$former_block_statement`. Since this expression originated from a block
statement though, it is a `Token::InternedStatement` and not a valid
expression.

I've updated the parser to accept interned statements in an expression
position, and we just later validate that they're either
`StatementKind::Expression` or `StatementKind::Semi` nodes.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.