fix: Fix canonicalization bug (#6033)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/6032

## Summary\*

This issue was because when we sort commutative terms we sorted them in
a BTreeSet, but this meant any identical terms were deduplicated. I
changed it to keep track of how many of each term there was so that we
don't accidentally change the expression when canonicalizing it.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.