feat: show test output when running via LSP (#6049)
# Description

## Problem

Alvaro told me that it would be useful to see the output of `println`
statements when running tests via LSP.

## Summary

Output is now always shown when running tests via LSP. If you are
running a single test it's likely you are debugging it too.

## Additional Context

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.