feat(metaprogramming): Add `#[use_callers_scope]` (#6050)
# Description

## Problem\*

We have many methods now such as `Quoted::as_type` or
`Quoted::as_trait_constraint` which resolve items in the current scope.
It is useful to change the scope these are resolved in but somewhat
impractical to do what we did with `Expression::resolve` and have them
all take module parameters.

## Summary\*

I've added `#[use_callers_scope]` which can be added to a comptime
function to change the scope any builtins resolve in to that of the
caller. This also works in helper functions and closures.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.