feat(devex): stop nargo from rebuilding every time
This could become quite annoying, as changing a .nr file would trigger
rebuild, which is clearly not what we want.

Added a colorful message to remind the user the behavior has changed.
Apologies, if this was not enough, and you were using nargo and
expecting that it still builds every time. It doesn't any more, unless
you explicitly tell it to.

(Also, the script starts the debug version. If you built release, you
can run it manually by specifying the path to the binary.)