feat: check unconstrained trait impl method matches (#6057)
# Description

## Problem

Fixes #5717

## Summary

In a previous PR we started allowing `unconstrained` and `comptime` in
trait impls. But two things remained:
1. Defining an `unconstrained` trait impl method would produce a
warning.
2. We should error if a trait method is unconstrained but the trait impl
method is not, or the other way around.

This PR will also include "unconstrained" in trait imp method stubs in
LSP.

## Additional Context

I initially also errored if there's a mismatch in `comptime`, but that's
probably not good as we have some implementation of `Append` be comptime
and some not. This made me wonder what's the purpose of `comptime` in
functions, if non-comptime functions can also be interpreted.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>