fix: Always parse all tokens from quoted token streams (#6064)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/5803
Resolves https://github.com/noir-lang/noir/issues/6059

## Summary\*

If we ever had a token stream with a subset of valid input we'd parse
just that subset instead of erroring that the full input doesn't parse.

## Additional Context

We don't have EOF tokens in token streams so I used the `end()`
primitive instead.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.