fix!: Infer globals to be u32 when used in a type (#6083)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/6081
Resolves https://github.com/noir-lang/noir/issues/6082

## Summary\*

Now that we check kinds explicitly we were erroring if globals weren't
already `u32` yet leading to a lot of existing code needing to be
changed.

If we change the exact equality to a unification check we can instead
infer the globals need to be a `u32` without requiring user code to be
changed.

This PR is still breaking since there were causes we didn't catch before
of invalid kinds being used in array lengths, e.g. `fn foo<let N: u8>(a:
[Field; N])` in one of our unit tests.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.