fix: Unify macro result type with actual type (#6086)
# Description

## Problem\*

Resolves an issue from @asterite on slack

## Summary\*

When we type check a macro call we don't know the actual type so we
return a fresh type variable.

When we later execute this code at comptime we do know the type, so we
should unify the type variable then, potentially catching any
inconsistencies. This also makes the value of the type variable known
(during evaluation, not type checking!) as in the new test case.

## Additional Context

Also fixed an issue where you couldn't call a method as a macro directly
and needed an extra `unquote!` call.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.