feat: let LSP suggest macro calls too (#6090)
# Description

## Problem

There was no way to get completions for macro calls, so the user had to
complete, go back and add the `!`.

## Summary

Now if something can be a macro call it's suggested too.

A special case is done for `unquote!`: it's only suggested as a macro
call, never as a function call.


![lsp-suggest-macros](https://github.com/user-attachments/assets/3bd6e38c-dfed-4391-b597-50feabef9539)

## Additional Context

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>