feat: (LSP) show global value on hover (#6097)
# Description

## Problem

I noticed Rust Analyzer shows a `const` value, so I figured we could do
the same thing with simple globals.

## Summary

This is only done for globals with "simple" types (bools, strings,
integers, arrays, tuples). It can be useful to avoid having to go to the
global definition to check the value, in case it's needed.


![lsp-globals](https://github.com/user-attachments/assets/8886e78d-f18c-4b2a-939b-c3f98c242044)

## Additional Context

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.