feat: Implement solver for mov_registers_to_registers (#6089)
# Description

## Problem\*

mov_registers_to_registers was a suboptimal codegen.

## Summary\*


Moving a set of registers to another set a registers with possible
overlap among them is an operation that is used in the codegen for
calling conventions. In order to solve that, in the past, we just made a
backup of every source register. That is inefficient and generates more
than the optimal number of register movements. In order to achieve the
optimal number of movements, this PR implements a solver that breaks the
loops (like MOV 2=>3 MOV 3=>2, but of arbitrary size) by adding one
temporary register per loop. We then perform the moves following a DFS
on the dependency chain after the loops have been broken.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.