feat(perf): Allow array set last uses optimization in return block of Brillig functions (#6119)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

Part of general effort to reduce Brillig bytecode sizes.

## Summary\*

We currently only perform the array set last uses optimization on ACIR
functions. This makes sense as the pass only looks only last uses in a
single block. We can start to expand the last uses opt to encompass more
than just ACIR functions and make it aware of the last array set of an
array across blocks. This is quite complex though so this PR currently
just attempts to look at the return block of Brillig functions and see
the effects.

## Additional Context


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.