feat: do not double error on import with error (#6131)
# Description

## Problem

In a code like this:

```noir
mod moo {
    struct Foo {}
}

use moo::Foo;

fn main() {}
```

you'd previously get two errors:
1. `moo::Foo` is private
2. `moo::Foo` is unused

There's no need to give two errors here: saying that `moo::Foo` is
private is enough and it doesn't matter that it's used or not because
it's not even accessible.

## Summary

This PR makes it so that only one error is given in this case.

## Additional Context

None.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>