fix: allow providing default implementations of unconstrained trait methods (#6138)
…ethods

# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We previously had a bug where it was not possible for us to provide a
default implementation for an unconstrained method on a trait as the
compiler would consider it as constrained. We now pass through the
runtime type properly so that these methods are considered
unconstrained.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.