fix: let token pretty printer handle `+=` and similar token sequences (#6135)
# Description

## Problem

Just a small thing: we have a token pretty printer and it was printing a
sequence of tokens like `+ =` like "+ =" instead of not putting a space
between them.

## Summary

Fixes the above.

## Additional Context

I'd add a test, but I get a linker error when trying to run tests on the
frontend project since a few weeks/months (I think it's because of
chumsky, not sure!). In any case this behavior is not critical, it's
just helpful for outputting quoted values at compile-time.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.