feat: simplify sha256 implementation (#6142)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This unconstrained helper looks like it's taken a constrained
implementation and not modified it to take advantage of the extra
flexibility of the unconstrained runtime. I've then changed it to use a
runtime for-loop to reduce the number of iterations.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.