test(plonky2): run plonky2_show_plonky2_regression test with relative instead of absolute paths
When running the tests, instead of passing the `--program-dir` parameter to
nargo, just change the current dir to the program dir. This causes the produced
line info comments in the `--show-plonky2` output to use relative, instead of
absolute paths for the source files, which makes the test data (expected plonky2
output) portable accross different computers, that use a different local
directory for the main git repo.