feat: allow silencing an unused variable defined via `let` (#6149)
# Description

## Problem

Resolves #6148

## Summary


![lsp-allow-unused](https://github.com/user-attachments/assets/738168ab-cfd6-468e-8abc-fbb0e30e60b6)

## Additional Context

This is only for `let`, which is what we currently need.

Should this be documented? I guess so, but I couldn't find where in our
docs we explain `let`. (on the other hand we have `deprecated` but it
seems to also be undocumented)

## Documentation

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.