fix: (LSP) make goto and hover work well for attributes (#6152)
# Description

## Problem

Hovering or using "Go to definition" on an attribute that was linked to
a function that generated code wasn't working well (it would jump to one
of the generated definitions instead of to the attribute function).

## Summary

Before (note how the docs are for the generated code, not the attr
function):


![lsp-hover-before](https://github.com/user-attachments/assets/2cb325c9-b73f-4f1c-a885-b842fd276c68)

After:


![lsp-hover-after](https://github.com/user-attachments/assets/d75a849e-8410-4e71-9d71-324c19714964)


## Additional Context

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.