chore(ci): Update gates diff action to not post Brillig sizes report with no changes (#6157)
…mment when not needed

# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

No issue as found while making the public functions size report on
aztec-packages.

## Summary\*

We currently repost the Brillig opcode report even when there have been
no changes. This updates the noir-gates-diff to the most recent commit
on the noir-gates-diff `main` branch which has the fix.


## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.